Refactor logging, show flights for aircraft option
