+++ src/commands/company.ts
@@ -1,16 +1,19 @@
 import yargs, { BuilderCallback, CommandModule } from 'yargs';
 import chalk from 'chalk';
 
-import { cliTable } from '../utils/cli-table';
 import { CommonConfig } from '../types/commonTypes';
 import { getCompany } from '../api/getCompany';
 import { getCompanyFleet } from '../api/getCompanyFleet';
 import { getCompanyFlights } from '../api/getCompanyFlights';
 import { getCompanyFbos } from '../api/getCompanyFbos';
 import { Company } from '../types/Company';
-import { Aircraft, aircraftStatuses } from '../types/Aircraft';
+import { Aircraft } from '../types/Aircraft';
 import { Flight } from '../types/Flight';
 import { Fbo } from '../types/Fbo';
+import { logFlights } from '../loggers/logFlights';
+import { logCompany } from '../loggers/logCompany';
+import { logCompanyFleet } from '../loggers/logCompanyFleet';
+import { logCompanyFbos } from '../loggers/logCompanyFbos';
 
 const log = console.log;
 
@@ -22,13 +25,14 @@ const builder = (yargs: yargs.Argv<CommonConfig>) => {
       choices: ['fleet', 'flights', 'fbos'],
     })
     .option('page', {
-      'describe': 'Page number (flights/fbos only)',
+      'describe': 'Page number (flights only)',
       'type': 'number',
+      'alias': 'p',
     })
     .example('$0 company','Get summary information for your company')
     .example('$0 company fleet','List your aircraft')
     .example('$0 company flights','List your flights')
-    .example('$0 company flights --page=2','List your flights, showing page 2')
+    .example('$0 company flights -p=2','List your flights, showing page 2')
     .example('$0 company fbos', 'List your FBOs');
 }
 
@@ -46,14 +50,14 @@ export const companyCommand: CompanyCommand = {
 
       if (typeof argv['action'] === 'undefined') {
         const company: Company = await getCompany(argv['companyId'], argv['apiKey'], argv['world']);
-        renderCompany(company);
+        logCompany(company);
       } else {
         switch (argv['action']) {
           case 'fleet': {
             const companyFleet: Aircraft[] = await getCompanyFleet(argv['companyId'], argv['apiKey'], argv['world']);
             if (companyFleet.length) {
               log(chalk.greenBright.bold('Your fleet of aircraft\n'));
-              renderCompanyFleet(companyFleet);
+              logCompanyFleet(companyFleet);
               log(`\nSuggested command: ${argv['$0']} aircraft <id>`);
             } else {
               log('Dude, where\'s your aircraft?! ' + chalk.magentaBright('✈'));
@@ -66,8 +70,8 @@ export const companyCommand: CompanyCommand = {
             const companyFlights: Flight[] = await getCompanyFlights(argv['companyId'], argv['apiKey'], argv['world'], page, limit);
             if (companyFlights.length) {
               log(chalk.greenBright.bold(`Your flights (Page ${page}, ${limit} per page)\n`));
-              renderCompanyFlights(companyFlights);
-              log(`\nSuggested command: ${argv['$0']} company ${argv['action']} --page=${page+1}`);
+              logFlights(companyFlights);
+              log(`\nSuggested command: ${argv['$0']} company ${argv['action']} -p=${page+1}`);
               log(`Suggested command: ${argv['$0']} airport <ICAO>`); 
               //log(`Suggested command: ${argv['$0']} flight <id>`); 
             } else {
@@ -79,7 +83,7 @@ export const companyCommand: CompanyCommand = {
             const companyFbos: Fbo[] = await getCompanyFbos(argv['companyId'], argv['apiKey'], argv['world']);
             if (companyFbos.length) {
               log(chalk.greenBright.bold('Your FBOs\n'));
-              renderCompanyFbos(companyFbos);
+              logCompanyFbos(companyFbos);
             } else {
               log('No FBO... no 100LL! ' + chalk.magentaBright('✈'))
             }
@@ -92,137 +96,4 @@ export const companyCommand: CompanyCommand = {
       console.error(chalk.bold.red(e.message))
     }
   }
-}
-
-const renderCompany = (company: Company): void => {
-  const infoTable = cliTable();
-  infoTable.push([chalk.green('Company'),company.Name,'','']);
-  infoTable.push([
-    chalk.green('Code'),
-    company.AirlineCode,
-    chalk.green('Reputation'),
-    Math.round(company.Reputation*1000)/10+'%',
-  ]);
-  infoTable.push([
-    chalk.green('Level'),
-    `${company.Level} (${company.LevelXP}xp)`,
-    chalk.green('Travel Tokens'),
-    company.TravelTokens
-  ]);
-  log(infoTable.toString()+'\n');
-
-  const dateTable = cliTable();
-  const lastConnection = new Date(Date.parse(company.LastConnection));
-  const lastReport = new Date(Date.parse(company.LastReportDate));
-  dateTable.push([chalk.green('Last Connection'), lastConnection.toLocaleString('en-GB')]);
-  dateTable.push([chalk.green('Last Report'), lastReport.toLocaleString('en-GB')]);
-  dateTable.push([chalk.green('UTC Offset'), company.UTCOffsetinHours + ' hours']);
-  dateTable.push([chalk.green('Paused'), company.Paused ? 'Yes' : 'No']);
-  log(dateTable.toString());
-}
-
-const renderCompanyFleet = (companyFleet: Aircraft[]): void => {
-  const aircraftTable = cliTable();
-  
-  aircraftTable.push([
-    chalk.green('Name'),
-    chalk.green('Ident'),
-    chalk.green('Airport'),
-    chalk.green('Status'),
-    chalk.green('ID')
-  ]);
-
-  companyFleet.forEach((Aircraft) => {
-    aircraftTable.push([
-      chalk.whiteBright(Aircraft.AircraftType.DisplayName),
-      Aircraft.Identifier,
-      Aircraft.CurrentAirport?.ICAO || '-',
-      aircraftStatuses[Aircraft.AircraftStatus],
-      Aircraft.Id
-    ])
-  });
-
-  log(aircraftTable.toString());
-}
-
-const renderCompanyFlights = (companyFlights: Flight[]): void => {
-  const flightTable = cliTable();
-
-  flightTable.push([
-    chalk.green('Ident'),
-    chalk.green('Airborne'),
-    chalk.green('=>'),
-    chalk.green('Landed'),
-    chalk.green('>='),
-    chalk.green('ID'),
-    chalk.green(''),
-  ]);
-
-  companyFlights.forEach((Flight) => {
-    const airborneTime = typeof Flight.AirborneTime !== 'undefined' ? 
-      new Date(Date.parse(Flight.AirborneTime)).toLocaleString('en-GB').substring(0,17) :
-      '-';
-
-    const landedTime = typeof Flight.LandedTime !== 'undefined' ? 
-      new Date(Date.parse(Flight.LandedTime)).toLocaleString('en-GB').substring(0,17) :
-      '-';
-
-    flightTable.push([
-      Flight.Aircraft.Identifier,
-      airborneTime,
-      Flight.DepartureAirport?.ICAO,
-      landedTime,
-      Flight.ArrivalActualAirport?.ICAO || '-',
-      Flight.Id,
-      Flight.RegisterState === 9 ? '✅' : '❌',
-    ])
-  });
-
-  log(flightTable.toString());
-}
-
-const renderCompanyFbos = (companyFbos: Fbo[]): void => {
-  const fboTable = cliTable();
-
-  fboTable.push([
-    chalk.green('Airport'),
-    chalk.green('Name'),
-    chalk.green('100LL'),
-    chalk.green('Sell'),
-    chalk.green('Jet'),
-    chalk.green('Sell'),
-    chalk.green('C'),
-    chalk.green('S'),
-    chalk.green('T'),
-    chalk.green('H')
-  ]);
-
-  companyFbos.forEach((Fbo) => {
-    fboTable.push([
-      Fbo.Airport.ICAO,
-      Fbo.Name,
-      Fbo.Fuel100LLQuantity + '/' + Fbo.Fuel100LLCapacity,
-      Fbo.AllowFuel100LLSelling ? `✅ ${Fbo.Fuel100LLSellPrice}` : '❌',
-      Fbo.FuelJetQuantity + '/' + Fbo.FuelJetCapacity,
-      Fbo.AllowFuelJetSelling ? `✅ ${Fbo.FuelJetSellPrice}` : '❌',
-      Fbo.CargoWeightCapacity,
-      Fbo.SleepingCapacity,
-      Fbo.AircraftTieDownCapacity,
-      Fbo.AircraftHangarCapacity
-    ])
-  });
-  
-  log(fboTable.toString());
-
-  log(
-    chalk.whiteBright.bold('\nKey ') + 
-    chalk.green('C') + 
-    ' Cargo capacity  ' + 
-    chalk.green('S') + 
-    ' Sleeping  ' + 
-    chalk.green('T') + 
-    ' Tiedowns  ' + 
-    chalk.green('H') + 
-    ' Hanger'
-  );
 }
